Split out gimli implementation into multiple files
Move object-file-parsing code into three separate files which can be
shared between different platforms.
