+++ src/symbolize/gimli/elf.rs
@@ -1,0 +1,114 @@
+use super::{Mapping, Path, Vec};
+use object::read::elf::{FileHeader, SectionHeader, SectionTable, Sym};
+use object::read::StringTable;
+use object::{Bytes, NativeEndian};
+
+#[cfg(target_pointer_width = "32")]
+type Elf = object::elf::FileHeader32<NativeEndian>;
+#[cfg(target_pointer_width = "64")]
+type Elf = object::elf::FileHeader64<NativeEndian>;
+
+impl Mapping {
+    pub fn new(path: &Path) -> Option<Mapping> {
+        let map = super::mmap(path)?;
+        let cx = super::cx(Object::parse(&map)?)?;
+        Some(mk!(Mapping { map, cx }))
+    }
+}
+
+struct ParsedSym {
+    address: u64,
+    size: u64,
+    name: u32,
+}
+
+pub struct Object<'a> {
+    /// Zero-sized type representing the native endianness.
+    ///
+    /// We could use a literal instead, but this helps ensure correctness.
+    endian: NativeEndian,
+    /// The entire file data.
+    data: Bytes<'a>,
+    sections: SectionTable<'a, Elf>,
+    strings: StringTable<'a>,
+    /// List of pre-parsed and sorted symbols by base address.
+    syms: Vec<ParsedSym>,
+}
+
+impl<'a> Object<'a> {
+    fn parse(data: &'a [u8]) -> Option<Object<'a>> {
+        let data = object::Bytes(data);
+        let elf = Elf::parse(data).ok()?;
+        let endian = elf.endian().ok()?;
+        let sections = elf.sections(endian, data).ok()?;
+        let mut syms = sections
+            .symbols(endian, data, object::elf::SHT_SYMTAB)
+            .ok()?;
+        if syms.is_empty() {
+            syms = sections
+                .symbols(endian, data, object::elf::SHT_DYNSYM)
+                .ok()?;
+        }
+        let strings = syms.strings();
+
+        let mut syms = syms
+            .iter()
+            // Only look at function/object symbols. This mirrors what
+            // libbacktrace does and in general we're only symbolicating
+            // function addresses in theory. Object symbols correspond
+            // to data, and maybe someone's crazy enough to have a
+            // function go into static data?
+            .filter(|sym| {
+                let st_type = sym.st_type();
+                st_type == object::elf::STT_FUNC || st_type == object::elf::STT_OBJECT
+            })
+            // skip anything that's in an undefined section header,
+            // since it means it's an imported function and we're only
+            // symbolicating with locally defined functions.
+            .filter(|sym| sym.st_shndx(endian) != object::elf::SHN_UNDEF)
+            .map(|sym| {
+                let address = sym.st_value(endian);
+                let size = sym.st_size(endian);
+                let name = sym.st_name(endian);
+                ParsedSym {
+                    address,
+                    size,
+                    name,
+                }
+            })
+            .collect::<Vec<_>>();
+        syms.sort_unstable_by_key(|s| s.address);
+        Some(Object {
+            endian,
+            data,
+            sections,
+            strings,
+            syms,
+        })
+    }
+
+    pub fn section(&self, name: &str) -> Option<&'a [u8]> {
+        Some(
+            self.sections
+                .section_by_name(self.endian, name.as_bytes())?
+                .1
+                .data(self.endian, self.data)
+                .ok()?
+                .0,
+        )
+    }
+
+    pub fn search_symtab<'b>(&'b self, addr: u64) -> Option<&'b [u8]> {
+        // Same sort of binary search as Windows above
+        let i = match self.syms.binary_search_by_key(&addr, |sym| sym.address) {
+            Ok(i) => i,
+            Err(i) => i.checked_sub(1)?,
+        };
+        let sym = self.syms.get(i)?;
+        if sym.address <= addr && addr <= sym.address + sym.size {
+            self.strings.get(sym.name).ok()
+        } else {
+            None
+        }
+    }
+}
